Support updating profile properties from a script
If you write a script to batch update user properties using the Engage()
method, then all the users will be updated to have their location set
to wherever the script was run from.

This change lets you write scripts that won't screw up the business metrics.

The functionality is isolated to a new method in order to keep the library
backwards compatible.